coord system and function draing works

+++ gui/src/gui_button/gui_abstract_button/gui_abstract_button_skin.cpp
@@ -26,7 +26,7 @@ gui::AbstractButtonSkin::~AbstractButtonSkin() = default;
 void gui::AbstractButtonSkin::Render() {
   m_render_texture.Resize(m_curr_texture_location.m_size);
   m_render_texture.Clear(glib::ColorRGBA(0, 0, 0, 0));
-  m_render_texture.CopyTexture(m_source_texture, glib::Vector2f(), m_curr_texture_location);
+  m_render_texture.CopyTexture(m_source_texture, glib::Vector2i(), m_curr_texture_location);
   m_render_texture.Display();
   m_texture = m_render_texture.GetTexture();
 }
